GetFieldData는 오류가 없었습니다. 주석을 달지 않은 제 잘못입니다. 씨앗 심기 시 그래픽의 변화 구현, 시간의 흐름에 따른 크기 변화 구현. 식물에 따른 밭의 그래픽을 따로 만들 필요가 없습니다. 지우고 싶다면 지우십시오. 종현 선배의 조언에 따라 UI가 비활성화 될 때 보이지 않은 곳으로 이동하는 것이 아니라 비활성화하는 것으로 변경하였습니다. 구현 로직에 대해선 GetFieldData에 달린 주석을 참고하시기 바랍니다. GetGrownPercent가 0 또는 1만을 반환하던 문제를 해결하였습니다.

+++ Assets/Script/Class/WPField.cs
@@ -81,7 +81,7 @@ public class WPField
 
 
         float percent = 
-            WPDateTime.CompareTime(WPDateTime.Now, startedTime) /
+            (float)WPDateTime.CompareTime(WPDateTime.Now, startedTime) /
             (Convert.ToInt32(seedData[seedIndex]["eGrowthTime"]) * 24);
 
         if (percent > 1f)
